fixed: got tests working
